+++ test/conformance/fn_ref/fail/non_nullptr_init.2.fail.cpp
@@ -1,4 +1,4 @@
-#include "../__constant_wrapper.hpp"
+#include "__constant_wrapper.hpp"
 #include "test_function.hpp"
 
 int main() {