+++ test/test_fallback_macros.hpp
@@ -31,4 +31,9 @@
 # define EBD_TEST_TRY_BUG__MSVC_1944_1
 #endif
 
+#if 0
+/// TODO: Following up <https://github.com/llvm/llvm-project/issues/106660>.
+# define EBD_TEST_TRY_BUG__Clang_106660
+#endif
+
 #endif // TEST_FALLBACK_MACRO_HPP_